requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
TensorFlow==v2.13.1
Flask-Cors==4.0.0
Pillow==10.1.0